Module Main.Main

This entire file is parameterised by an interface as specified in the InOut Module.

Parameters

IO : Blocklib.InOut.T

Signature

val webpage_link : string

val webpage_issues : string

val webpage_icons : string

val urltag_lang : string

URL tags

val urltag_path : string

val trace : string list Stdlib.ref

A trace of the menu function to help debugging.

val add_trace : string -> unit

Adding a message in the trace

val get_trace : 'a -> string list

Get the full trace.

val errorTranslationsDefault : string * string * string * string

The default error messages.

val errorTranslations : (string * string * string * string) Stdlib.ref

The translations needed to print error messages.

val file_signature : 'a -> string

val get_translations : 'a -> ((string * string, string) PMap.t * string list) Lwt.t

Getting and parsing the translations file.

val main : unit Lwt.t

The main script.