Module Tujkuko.Units

type t

The type of units. Note that it might include some metric prefix.

type value = float * t

A value with a unit.

val parse : string -> t option

Parse a unit from a string, retuning None in case no unit is recognised. This function follows the standard notations (for instance, min for minutes, and not m.

val parse_list : string -> t list

Parse a unit from a user string. This is less strict than parse and will accept m for minutes (although meters will also be returned). The first element of the list is the prefered one (typically the one returned by parse).

val print : t -> string

Print the unit.

type kind =

| Mass

| Length

| Volume

| Time

| Temperature

The kind of the unit.

val get_kind : t -> kind

Get the kind of a unit.

type system

A unit system (e.g. metric)

val system_name : system -> string

Get a name for a system.

val parse_system : string -> system option

Inverse of system_name.

val get_system_kind : system -> kind

Each system is only made for a specific kind of units.

val get_base_unit : system -> t

Given a system, return its base unit.

val get_system : t -> system

Given a unit, return its associated system.

val system_units : system -> t list

All the units defined in a system.

val all_systems : (kind, system list) PMap.t

All the available systems, for each unit kinds. The first one is usually the most preferable one.

exception InvalidKind

val normalise : system -> value -> value

Given a unit system, normalise a value. Raises InvalidKind if the system kind does not correspond to the given unit kind.

val self_normalise : value -> value

Same as normalise, but uses the value’s system.

val convert : t -> value -> float

Given a unit and a value, convert the value to this particular unit. Raises InvalidKind if the value and the unit do not correspond to the same unit kind.

val increase_unit_value : value -> value option

Given a value and a unit, use a higher unit to represent the value. Return None if there is no higher avalaible units in the current unit system.

val decrease_unit_value : value -> value option

Given a value and a unit, use a lower unit to represent the value. Return None if there is no lower avalaible units in the current unit system.