Module Tujkuko__Import

Importing recipes.

val from_json : string -> Tujkuko.Recipe.t

Import recipes from a JSON file. May raise Invalid_arg.

val import_translations : string -> string -> (string * string, string) PMap.t * string list

import_translations fileName fileContent reads fileContent as a JSON object representing translations in different languages. It then returns a translation object (associating a pair of identifier and language code to its translation) as well as the list of languages codes found in the file.