Module Tujkuko__Interpolation

Define how each measurments change the global factor.

type t

The main type describing how a given unit interpolate with the global factor.

val constant : t

The “correlation” of a constant measurment, that doesn’t change with the global factor.

val linear : t

A linear factor, where the global factor corresponds exactly to the given unit.

val default : Tujkuko.Units.kind -> t

The default interpolation factor for a unit kind.

val export : t -> Yojson.Safe.t

Functions to import and export correlations. The importation function may raise Invalid_arg.

val import : Yojson.Safe.t -> t