Module Tujkuko__Recipe

Define the data structures to store recipes.

type language = string

Each language is identified by its ISO 639 code.

type item =

| Sentence of string
A string that will be displayed as-is.

| Unit of float * float * Tujkuko.Interpolation.t * Tujkuko.Units.t option
An expected quantity, with a minimum and maximal values (which can be equal), the interpolation, and an optional unit.

A step is a list of items.

type step = item list

type info = {

id : string;
An identifier, unique amongst siblings.

picture : string option;
An optional picture of the step.

description : (language, step) PMap.t;
For each language, a description of the step.

hints : (language, step list) PMap.t;
For each language, a list of hints.

}

Information that can be stored at any point of the step.

type t =

| End

| Step of (info * t) list

The main type: a tree with all the recipes. The constructor End states that the recipe is finished. This is different than Step [] in the sense that no further step are meant to be added afterwards.